[PATCH] Define PAGE_SIZE for aarch64
If this page size isn't set statically, the page size is instead loaded
from libc which calculates it when the process is initialised. If muslc
is not providing the runtime, the PAGE_SIZE becomes 0.

This can lead to a strange issue where malloc will assume that every
call to sbrk fails (including the first) and will always fall back to
mmap. For some reason, this ends up confusing malloc after an initial
allocation of 0 bytes beginning at the _end_ of the static morecore
region.

Co-authored-by: Curtis Millar <curtis@curtism.me>
Signed-off-by: alwin-joshy <joshyalwin@gmail.com>